+++ benchmarks/binned.py
@@ -48,6 +48,23 @@ class Binned2dConcat:
         self.da.bins.concat('x')
 
 
+class Binned2dConcatInner:
+    """
+    Benchmark reduction with 'concat' along inner for 2d binned data
+    """
+    params = (list(2**np.arange(10, 16)), )
+    param_names = ['nbin']
+    timeout = 300.0
+
+    def setup(self, nbin):
+        binned = sc.data.binned_x(nevent=2 * nbin, nbin=nbin)
+        y = sc.linspace(dim='y', start=0, stop=1, num=2, unit='m')
+        self.da = sc.bin(binned, edges=[y])
+
+    def time_bins_concat_long_outer(self, nbin):
+        self.da.bins.concat('y')
+
+
 class Lookup:
     """
     Benchmark map operations using sc.lookup()